Backend/requirements.txt
blinker==1.7.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pdfminer.six==20231228
pycparser==2.22
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
openai
langchain
python-dotenv
